#NTR - Fix plugin model orphan removal

+++ engine/Shopware/Components/Plugin/Namespace.php
@@ -332,13 +332,15 @@ class Shopware_Components_Plugin_Namespace extends Enlight_Plugin_Namespace_Conf
             /** @var $em Shopware\Components\Model\ModelManager */
             $em = $this->Application()->Models();
 
-            $form = $plugin->Form();
-            if($form->getId()) {
-                $em->remove($form);
-            } else {
-                $em->detach($form);
+            if($plugin->hasForm()) {
+                $form = $plugin->Form();
+                if($form->getId()) {
+                    $em->remove($form);
+                } else {
+                    $em->detach($form);
+                }
+                $em->flush();
             }
-            $em->flush();
 
             $query = 'DELETE FROM Shopware\Models\Menu\Menu m WHERE m.pluginId = ?0';
             $query = $em->createQuery($query);
